Move UpdatePrice to its own component
