+++ src/components/PlantCard.js
@@ -1,49 +1,12 @@
-import React, { useRef, useState } from "react"
+import React, { useState } from "react"
 import DeleteButton from "./DeleteButton"
 import InStockButton from "./InStockButton"
+import UpdatePriceButton from "./UpdatePriceButton"
 
 function PlantCard({ plants, setPlants, plant, onDeleteItem }) {
-  const { id } = plant
   const [isInStock, setIsInStock] = useState(true)
   const [showForm, setShowForm] = useState(false)
   const [newPrice, setNewPrice] = useState(0)
-  const formRef = useRef(null)
-
-  const showUpdatePriceForm = () => {
-    setShowForm(!showForm)
-  }
-
-  function handleChangePrice(event) {
-    setNewPrice(event.target.value)
-  }
-
-  function updatePrice(data) {
-    let updatedPlants = plants.map((plant) =>
-      plant.id === data.id ? { ...plant, price: data.price } : plant
-    )
-    setPlants(updatedPlants)
-  }
-
-  function onSubmitPrice(event) {
-    event.preventDefault()
-    fetch(`http://localhost:6001/plants/${id}`, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        price: newPrice,
-      }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        updatePrice(data)
-        if (formRef.current) {
-          formRef.current.reset()
-        }
-        setShowForm(false)
-      })
-  }
 
   return (
     <>
@@ -51,26 +14,15 @@ function PlantCard({ plants, setPlants, plant, onDeleteItem }) {
         <div className="card-container">
           <DeleteButton onDeleteItem={onDeleteItem} plant={plant} />
           <InStockButton isInStock={isInStock} setIsInStock={setIsInStock} />
-          <button className="secondary" onClick={showUpdatePriceForm}>
-            Update Price
-          </button>
-          {showForm && (
-            <div className="update-price-form">
-              <form onSubmit={onSubmitPrice} ref={formRef}>
-                <label htmlFor="price">New Price</label>
-                <br />
-                <input
-                  className="update-price-input"
-                  type="text"
-                  id="price"
-                  onChange={handleChangePrice}
-                />
-                <button type="submit" value="Submit">
-                  Submit
-                </button>
-              </form>
-            </div>
-          )}
+          <UpdatePriceButton
+            plant={plant}
+            showForm={showForm}
+            setShowForm={setShowForm}
+            newPrice={newPrice}
+            setNewPrice={setNewPrice}
+            plants={plants}
+            setPlants={setPlants}
+          />
         </div>
       </li>
     </>
